Revert "Autoremove packages no longer required"
This reverts commit 256f9e654e9a06a10a865eecab67c4812295f769.

This change is making it impossible for our current packer based systems
to build new images. There was no stated reasoning as to the autoremove
was added, just that it was and it made an arbitrary upgrade requirement
without vetting if it was even possible to make the move in our current
environments.

Change-Id: If3bcce8f2e0c285609b0efbde293d4e1bf53b787
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>